Remove duplicates from sorted array
Type Inference: Here, auto allows the compiler to determine that last should be of type vector<int>::iterator (or std::vector<int>::iterator), which is the return type of the unique algorithm. This is especially useful for iterators, which can be verbose to write out explicitly.

Sorting: Necessary before using unique to ensure all duplicates are adjacent.

unique Algorithm: Moves the unique elements to the front and returns an iterator to the new end of the unique range.

erase Method: Removes elements from the vector starting from the iterator returned by unique to the end of the vector.